pyproject.toml
[project]
name = "customer-support-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.46.0",
    "langchain-google-genai>=3.0.0",
    "langgraph>=1.0.1",
    "mem0ai>=1.0.0",
    "python-dotenv>=1.2.1",
]
